[Refactor] Use the arg-parser command line arguments
This refactors the tool to use the `arg-parser` package
from the `stanza` library. This makes for a consistent commandline
interface across the stanza project and allows for easier ability
to update and configure the tool as we go on.

I had to change the way that `build` and `repl` work. Previously,
the remaining arguments after the command were applied to the
stanza invocation. This unfortunately is degenerate for the case
where the `build` command has arguments. How do you decide what
goes to slm and what goes to stanza.

Instead, now you can use environment variables to set the
additional build arguments for `stanza` invokations.