Add Build Target and Fix Bugs (#38)
* Removed unused function and simplified the error logging.

* Removed trailing whitespace

* Added target for build command

Closes #36

This allows slm to build other targets which will be helpful for
building unit tests, among other things.

* [Bugfix] Updated logging to use proper access to stdout/stderr

The previous implementation was accessing `STANDARD-OUTPUT-STREAM` and
`STANDARD-ERROR-STREAM` directly. This makes it impossible to write
unit tests that cover output generated to stdout/stderr because it
is not interceptable.

By using `current-output-stream()` and `current-error-stream()` - we
can write unit tests that check the text output of the functions.