[Refactor] Changed stanza.proj tree structure
Closes JITX-7191

This commit changes the way that the `stanza.proj` files are created for
projects.

In the old way, the `.slm/stanza.proj` was effectively the `root` stanza
project file. This meant that to build, slm had to change directories to `.slm`
and then run the `stanza build` there.

This caused a lot of complexity. It meant that I couldn't just run
`stanza build` at the root directory. It also broke dependency resolution
in vscode via the stanza language server.

This commit creates the root `stanza.proj` at the root of the project
and then has the root `stanza.proj` import `.slm/stanza.proj`. The
`.slm/stanza.proj` still manages all of the resolved dependencies of
the project and abstracts that away from the end user. With this setup:

1.  The user can run `stanza build` from the root
2.  VSCode language server continues to work as expected.
3.  The structure of the project is a lot more clear.
4.  We can more easily define configuration options that affect the build
    without hiding them inside the `slm` executable. Specifically - I
    moved the `-pkg pkgs` command line option and put it  in the `stanza.proj`
    file for the `main` target. This means `stanza build` will just work.