[Bugfix] Dependency resolution for grandchildren
Closes JITX-7192

This was a bug because the `False` case and the `PathDependency` case were
being lumped together.

The `PathDependency` case for grandchildren is not handle-able in the
current implementation. This just commit just stops the bleeding and
makes the user more aware of the shortcoming instead of a random
'Key Not Found' error.