Added check for `git` before generating sub-processes
This commit modifies the `init/build/repl/publish` commands to
first check that a `git` binary exists before attempting to
launch `git` sub-processes. This will allow the `slm init`
command to gracefully degrade when not `git` repository is
present.
For the `build/repl` commands - if the `slm.toml` or any of
the derived dependencies expect to use `git` to clone a dependency,
then the commands will now error out.

The `publish` command checks for `git` first and errors out if
not found.