[Bugfix] Fix `.slm/stanza.proj` include for dependencies
This causes the `include .slm/stanza.proj` to be optional by default.
This is necessary so that when we include a dependency that also
builds as a binary to our project, that dependency won't fail to build
looking for a `dep/.slm/stanza.proj` that doesn't exist.

I've added this to the `slm` stanza.proj as well because there is a future
where we want to use `slm` as both a binary and a library.