[Bugfix] Added error handling for when `git ls-remote` fails.
If the user tries to access a repo that doesn't exist, the old code
would just not complain and carry on as if no error had occured.
For the publish command - that was not good but at least unlikely
to cause harm. With the new `add` command, I need to know if
`ls-remote` fails.