Updated the 'clean' to clear our cached files.
Previous, the default code for a project was relying on the `pkg-cache` in the
stanza distribution for building the project. This was causing the
`#env-var(SLM_BUILD_VERSION)` code not to work well. The caching was causing
the version number from previous builds to override because the cache didn't think
that anything had changed.

This commit causes new projects to use the `pkg-cache:` statement in the `stanza.proj`
file to point at `.slm/pkg-cache`. This causes the hashed pkgs to be looked for in the
local package cache instead of in `stanza/pkg-cache`.

The `clean` command can now remove all of the contents of the `.slm/pkg-cache` directory
and cause the project to start from a clean slate.

Note:
You may need to clear out any previous compiled/hashed packages in `stanza/pkg-cache`
for this new method to work.