[Refactor] Moved git/path dependency definitions to separate files.
The `dependency.stanza` file had a lot of different content in it.
With the introduction of the `TaskDependency` it made sense to use
a separate file.

I've now broken the `GitDepedency` and `PathDependency` into
separate files for similar benefits. I'm going to do some
further refactors shortly for `GitDependency` related code.