Added TaskDependency Support and some Refactoring
This refactors some of the `dependencies.stanza` file to move
`lock` related code to a separate file. I then add a unit test for
the lock code that generates `slm.lock`.

This is a rough cut. The `slm.lock` related resolution code has
a ton of bugs in it - mostly relatd to not handling grand-children
and great-grand-children dependencies well.

More work to come.