Added new environment variables for differentiating build vs repl
This adds a new env-var defined in both build and repl that
indicates whether we are building the dynamic library case or not.
This is separate from the existing env-vars `SLM_BUILD_STATIC` and
`SLM_BUILD_SHARED`.

This case primarily makes it easier to use conan from the task
dependency case.