Updated the init command to better deal with git repos
Bhusan reported that the `slm init` was not correctly detecting
that a git repo already exists when when from a sub-directory
of a git repo. This was because it was only looking for `.git`
directory which is not sufficient.

This now uses `git status` to determine if we are in a repo
or not.