[Bugfix] Fixed the `add` command mangling the path dependencies
Previously - I had added the `env-sub-enable` argument that made it so when parsing the slm toml file it didn't substitute the env vars for any of the values. This was necessary to support the `add` for path projects - specifically in JITX.

Somehow one of the functions that had this argument wasn't passing it on to the base class.