Added the `SLM_DIR` environment variable.
In various `stanza.proj` files, I often encounter a pattern where I
want to reference the top level `.slm` directory of a project. This
ends up being a relative path reference with a `..` in it.

This poses a problem if you want to run the project locally to build, for example,
the unit tests for a project.

This env-var makes it so the stanza.proj files don't have to use the relative
path and can instead reference a direct location with `{SLM_DIR/deps/some-lib/...`